+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Ability for computed property assertions to accept regular expressions
 - Displaying computed selectors in assertion errors
 - `checked` property and property creator
+- `selected` property and property creator
 
 ### Fixed